+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/platform/server/api/authorization/auth/AuthServiceNone.scala
@@ -5,7 +5,7 @@ package com.digitalasset.platform.server.api.authorization.auth
 
 import java.util.concurrent.{CompletableFuture, CompletionStage}
 
-import com.daml.ledger.participant.state.v1.{AuthService, Claims}
+import com.digitalasset.ledger.api.auth.{AuthService, Claims}
 import io.grpc.Metadata
 
 /** An AuthService that rejects all calls by always returning an empty set of [[Claims]] */
