+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/platform/server/api/authorization/services/PartyManagementServiceAuthorization.scala
@@ -3,7 +3,7 @@
 
 package com.digitalasset.platform.server.api.authorization.services
 
-import com.daml.ledger.participant.state.v1.AuthService
+import com.digitalasset.ledger.api.auth.AuthService
 import com.digitalasset.grpc.adapter.utils.DirectExecutionContext
 import com.digitalasset.ledger.api.v1.admin.party_management_service.{
   AllocatePartyRequest,
