+++ ledger/ledger-api-integration-tests/BUILD.bazel
@@ -44,6 +44,7 @@ dependencies = [
     "//ledger-api/rs-grpc-akka",
     "//ledger-api/rs-grpc-bridge",
     "//ledger-api/testing-utils",
+    "//ledger/ledger-api-auth:ledger-api-auth",
     "//ledger/ledger-api-client:ledger-api-client",
     "//ledger/ledger-api-common:ledger-api-common",
     "//ledger/ledger-api-common:ledger-api-common-scala-tests-lib",
